fix: Copilotレビュー指摘を修正（SSRF対策・アーキテクチャ整理）
- image-proxy: URLをHTTPS限定かつYahoo Japan CDN（*.yimg.jp等）の
  ホワイトリストに限定してSSRF/オープンプロキシ脆弱性を修正
- NewItemPage・EditItemPage をsrc/components/pages/に抽出し、
  ルートファイルを薄いラッパーのみに整理してアーキテクチャ仕様に準拠
